ws-products/src/main/java/com/rafaeldbl/wsproducts/service/CategoryService.java
package com.rafaeldbl.wsproducts.service;

import com.rafaeldbl.wsproducts.model.Category;
import com.rafaeldbl.wsproducts.repository.CategoryRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service("ICategoryService")
public class CategoryService implements ICategoryService {

    private final CategoryRepository repository;

    @Autowired
    public CategoryService(CategoryRepository repository) {
        this.repository = repository;
    }

    @Override
    public List<Category> getAllCategories() {
        return repository.findAll();
    }
}
